Banner updates

+++ src/events/message.js
@@ -28,24 +28,26 @@ module.exports.run = ({ client, serverInfo, message, args, sql, config, sendEmbe
     }
 
     //* Links filter
-    if (!message.member.isCH && !message.member.roles.has(serverInfo.roles.linksFiles)) {
-        if (!(config.permits[message.author.id] && config.permits[message.author.id].channel === message.channel.id && config.permits[message.author.id].until > new Date().getTime())) {
-            if (!config.whitelistedLinksChannel.includes(message.channel.id)) {
-                let matches = message.content.match(new RegExp(/https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/ig));
-                let filtered = matches ? matches.filter(m => {
-                    if (m.endsWith("discord.gg/alphaconsole")  ||
-                        m.includes("imgur.com") ||
-                        m.includes("reddit.com") ||
-                        m.includes("gyazo.com") ||
-                        m.includes("prntscr.com"))
-                        return false
-                    else
-                        return true
-                }) : null;
-
-                if (filtered && filtered[0]) {
-                    message.delete().catch(e => { });
-                    sendEmbed(message.author, "Links are not allowed in this channel.")
+    if (message.channel.id !== serverInfo.channels.setBanner) {
+        if (!message.member.isCH && !message.member.roles.has(serverInfo.roles.linksFiles)) {
+            if (!(config.permits[message.author.id] && config.permits[message.author.id].channel === message.channel.id && config.permits[message.author.id].until > new Date().getTime())) {
+                if (!config.whitelistedLinksChannel.includes(message.channel.id)) {
+                    let matches = message.content.match(new RegExp(/https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/ig));
+                    let filtered = matches ? matches.filter(m => {
+                        if (m.endsWith("discord.gg/alphaconsole")  ||
+                            m.includes("imgur.com") ||
+                            m.includes("reddit.com") ||
+                            m.includes("gyazo.com") ||
+                            m.includes("prntscr.com"))
+                            return false
+                        else
+                            return true
+                    }) : null;
+    
+                    if (filtered && filtered[0]) {
+                        message.delete().catch(e => { });
+                        sendEmbed(message.author, "Links are not allowed in this channel.")
+                    }
                 }
             }
         }
@@ -299,6 +301,7 @@ function AutoResponseChannel(channelID, channels) {
 	if (channelID === channels.serverlog) return true;
 	if (channelID === channels.setSpecialTitle) return true;
 	if (channelID === channels.setTitle) return true;
+	if (channelID === channels.setBanner) return true;
 	if (channelID === channels.showcase) return true;
 	if (channelID === channels.suggestion) return true;
 	if (channelID === channels.staff) return true;
